+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Bugfixes
 
 - Fixes error when deleting multiple targets of the same entity (#500)
+- Fixes missing relation events on archetype cleanup (deleted target entities) (#501)
 
 ## [[v0.8.0]](https://github.com/mlange-42/ark/compare/v0.7.1...v0.8.0)